+++ rust/src/api/descriptor.rs
@@ -18,8 +18,7 @@ impl Descriptor {
         let signer: SwSigner = SwSigner::new(&mnemonic, is_mainnet)?.into();
         let script_variant = lwk_common::Singlesig::Wpkh;
         let blinding_variant = lwk_common::DescriptorBlindingKey::Slip77;
-        let desc_str =
-            lwk_common::singlesig_desc(&signer, script_variant, blinding_variant, is_mainnet)?;
+        let desc_str = lwk_common::singlesig_desc(&signer, script_variant, blinding_variant)?;
         Ok(Descriptor {
             ct_descriptor: desc_str.to_string(),
         })